+++ fiddles/angular/fiddle-0003-PhoneCatIntro/setup.sh
@@ -1,0 +1,23 @@
+#!/usr/bin/env bash
+
+source script.sh
+
+function catch() {
+    case $1 in
+        0)  echo "setup succeeded"
+            ;;
+        1)  echo "script.sh: failed @ clonePhoneCatRepo@14()";
+            ;;
+        2)  echo "script.sh: failed @ installPhoneCatRepo@14()";
+            ;;
+        *)  echo "fubar! Something went wrong."
+            ;;
+    esac
+    exit $1
+}
+# try
+(
+    clonePhoneCatRepo@14 || exit 1;
+    installPhoneCatRepo@14 || exit 2;
+)
+catch $?;