+++ fiddles/bash/fiddle-0061-InstallNpmCheckUpdates/test.sh
@@ -1,0 +1,23 @@
+#!/usr/bin/env bash
+
+source script.sh
+
+function catch() {
+    case $1 in
+        0)  echo "all tests succeeded"
+            ;;
+        1)  echo "script.sh: installNpmCheckUpdates() failed";
+            ;;
+        2)  echo "script.sh: checkForBowerUpdates() failed";
+            ;;
+        *)  echo "fubar! Something went wrong."
+            ;;
+    esac
+    exit $1
+}
+# try
+(
+    installNpmCheckUpdates || exit 1;
+    checkForBowerUpdates || exit 2;
+)
+catch $?;