+++ scripts/bin/house-node-fs-stop.sh
@@ -26,8 +26,6 @@ _pid=$(lsof -i:${_port} -t || echo '-1')  # Capture Process Id
   # Kill the process
   $(lsof -i:${_port} -t | xargs kill;) || exit 86
 
-  e
-
 )
 #catch
 _rc=$?